Created PostPage component for modularity.
